Add a list of providers to the provider attribute
Formerly, this sort of thing was done in the Puppet code by overriding the doc
method for the provider parameter, but now we can do it in the formatter! This
gives us freedom to turn those into magic links, etc. without tying the Puppet
code base to the header IDs of our formatted templates, etc.