Refactor reference printer to use new data schema
Using a bunch of collects, although not quite going all the way. I'm kind of
skeptical about the utility of this. Also, is this a good candidate for a
template, maybe? Also also, consider the utility of the various other Markdown
auto-formatter functions we left behind in the docs module.