(PUP-3826) Update API docs for Puppet 4 URL changes
For Puppet 4, a number of changes have been made to Puppet's API. These
include splitting it into a master and CA API, complete with separate url
prefixes and explicit versioning, and moving the environment out of the path
and into a query parameter. This commit updates the API docs accordingly.